refactor: reorganize service interfaces and utilities in rqueue-core and rqueue-web
Move backend-agnostic service interfaces to rqueue-core, ensuring clean separation between core abstractions and the dashboard UI layer:

**Service Interfaces Relocated:**
- Dashboard services → rqueue-core/src/main/java/com/github/sonus21/rqueue/web/
  - RqueueSystemManagerService
  - RqueueQDetailService
  - RqueueJobService
  - RqueueDashboardChartService
  - RqueueViewControllerService

- Metrics service → rqueue-core/src/main/java/com/github/sonus21/rqueue/metrics/
  - RqueueMetricsAggregatorService (renamed from RqueueJobMetricsAggregatorService)

**Package Organization Improvements:**
- Move Pebble template functions from rqueue-web/utils/pebble/ to rqueue-web/web/pebble/
  - DateTimeFunction, DeadLetterQueuesFunction, DefaultFunction, DurationFunction
  - ReadableDateTimeFunction, ResourceLoader, RqueuePebbleExtension
  - Aligns template utilities with web package structure

**Impact:**
- Backend modules (rqueue-redis, rqueue-nats) now depend only on rqueue-core
- rqueue-web continues to depend on rqueue-core and hosts service implementations
- Spring integration modules (rqueue-spring, rqueue-spring-boot-starter) keep rqueue-web dependency
- Improved package organization with utilities co-located with the packages they serve
