+++ rqueue-web/src/main/java/com/github/sonus21/rqueue/web/pebble/DateTimeFunction.java
@@ -14,39 +14,29 @@
  *
  */
 
-package com.github.sonus21.rqueue.utils.pebble;
+package com.github.sonus21.rqueue.web.pebble;
 
-import com.github.sonus21.rqueue.models.db.DeadLetterQueue;
-import com.github.sonus21.rqueue.utils.Constants;
+import com.github.sonus21.rqueue.utils.DateTimeUtils;
 import io.pebbletemplates.pebble.extension.Function;
 import io.pebbletemplates.pebble.template.EvaluationContext;
 import io.pebbletemplates.pebble.template.PebbleTemplate;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-import org.springframework.util.CollectionUtils;
 
-@SuppressWarnings("unchecked")
-public class DeadLetterQueuesFunction implements Function {
+public class DateTimeFunction implements Function {
 
-  public static final String FUNCTION_NAME = "dlq";
+  public static final String FUNCTION_NAME = "time";
 
   @Override
   public Object execute(
       Map<String, Object> args, PebbleTemplate self, EvaluationContext context, int lineNumber) {
-    List<DeadLetterQueue> deadLetterQueues = (List<DeadLetterQueue>) args.get("queues");
-
-    if (CollectionUtils.isEmpty(deadLetterQueues)) {
-      return "";
-    }
-    List<String> queues =
-        deadLetterQueues.stream().map(DeadLetterQueue::getName).collect(Collectors.toList());
-    return String.join(Constants.Comma, queues);
+    Long milli = (Long) args.get("milli");
+    return DateTimeUtils.formatMilliToString(milli);
   }
 
   @Override
   public List<String> getArgumentNames() {
-    return Collections.singletonList("queues");
+    return Collections.singletonList("milli");
   }
 }
